pyproject.toml
[project]
name = "todo-service"
version = "0.1.0"
description = ""
authors = [
    {name = "ad",email = "antosha_klepikov@bk.ru"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi[all] (>=0.135.0,<0.136.0)",
    "sqlalchemy (>=2.0.47,<3.0.0)",
    "black (>=26.1.0,<27.0.0)",
    "uvicorn[standard] (>=0.41.0,<0.42.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false